# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ChatGPT Monitor API
  version: 0.6.0
  description: STEP-07 API contract. Authentication uses same-origin Secure cookies; credentials, OAuth device codes, channel secrets, and raw upstream evidence are never returned. Notification channels are configuration-only and remain disconnected.
  license:
    name: Proprietary
servers:
  - url: https://127.0.0.1:8080
paths:
...
  /api/v1/monitor/accounts:
    get:
      operationId: listAllocationMonitorAccounts
      summary: List monitor-owned accounts for allocation-service pull synchronization
      security:
        - allocationServiceBearer: []
      responses:
        '200':
          description: Read-only monitor account list without credential material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocationAccountListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/monitor/accounts/import-for-allocation:
    post:
      operationId: importAccountForAllocation
      summary: Import one credential for allocation-service monitoring ownership
      security:
        - allocationServiceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocationImportRequest'
      responses:
        '201':
          description: Account imported by the monitor service without returning credential material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocationAccountStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '503':
          $ref: '#/components/responses/ProviderUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/monitor/accounts/batch-status:
    post:
      operationId: getAllocationBatchStatus
      summary: Return item-level monitor status for 1-100 provider account ids
      security:
        - allocationServiceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocationBatchStatusRequest'
      responses:
        '200':
          description: Batch status with item-level not_found errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocationBatchStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/monitor/accounts/{providerAccountId}/status:
    parameters:
      - $ref: '#/components/parameters/ProviderAccountID'
    get:
      operationId: getAllocationAccountStatus
      summary: Return monitor status for one provider account id
      security:
        - allocationServiceBearer: []
      responses:
        '200':
          description: Account status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocationAccountStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-session
    allocationServiceBearer:
      type: http
      scheme: bearer
...
  parameters:
...
    ProviderAccountID:
      name: providerAccountId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 256
  responses:
...
    Unauthorized:
      description: A valid, non-revoked session is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationFailed:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: The active account does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The provider account is already active
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: Content-Type must be application/json
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ProviderUnavailable:
      description: The upstream credential provider returned a retryable failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Request could not be processed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    AllocationImportRequest:
      type: object
      additionalProperties: false
      required: [token, token_type]
      properties:
        label:
          type: string
          maxLength: 256
        token:
          type: string
          minLength: 1
          maxLength: 16384
          writeOnly: true
        token_type:
          type: string
          enum: [access_token, refresh_token, session_token, access, refresh, session]
    AllocationBatchStatusRequest:
      type: object
      additionalProperties: false
      required: [provider_account_ids]
      properties:
        provider_account_ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 256
    AllocationAccountStatus:
      type: object
      additionalProperties: false
      required: [provider_account_id, status, plan, auth_expiry, subscription_expiry, checked_at]
      properties:
        provider_account_id:
          type: string
          minLength: 1
          readOnly: true
        email:
          type: string
          format: email
          nullable: true
          readOnly: true
        status:
          type: string
          enum: [alive, dead_normal, dead_banned]
          readOnly: true
        plan:
          type: string
          enum: [free, plus, team, unknown]
          readOnly: true
        auth_expiry:
          type: string
          format: date-time
          readOnly: true
        subscription_expiry:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        checked_at:
          type: string
          format: date-time
          readOnly: true
    AllocationAccountListResponse:
      type: object
      additionalProperties: false
      required: [accounts]
      properties:
        accounts:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/AllocationAccountListItem'
        request_id:
          type: string
          readOnly: true
    AllocationAccountListItem:
      type: object
      additionalProperties: false
      required: [provider_account_id, status, plan, auth_expiry]
      properties:
        provider_account_id:
          type: string
          minLength: 1
          readOnly: true
        email:
          type: string
          format: email
          nullable: true
          readOnly: true
        status:
          type: string
          enum: [alive, dead_normal, dead_banned]
          readOnly: true
        plan:
          type: string
          enum: [free, plus, team, unknown]
          readOnly: true
        auth_expiry:
          type: string
          format: date-time
          readOnly: true
    AllocationBatchStatusResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AllocationBatchStatusItem'
    AllocationBatchStatusItem:
      type: object
      additionalProperties: false
      required: [provider_account_id]
      properties:
        provider_account_id:
          type: string
          minLength: 1
        status:
          type: string
          enum: [alive, dead_normal, dead_banned]
        plan:
          type: string
          enum: [free, plus, team, unknown]
        subscription_expiry:
          type: string
          format: date-time
          nullable: true
        checked_at:
          type: string
          format: date-time
        error:
          $ref: '#/components/schemas/AllocationItemError'
    AllocationItemError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum: [not_found, validation_failed]
        message:
          type: string
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
        request_id:
          type: string
        run_id:
          type: string